requirements.txt
amqp==2.5.0
billiard==3.6.0.0
boto3==1.9.209
botocore==1.12.209
cachetools==4.0.0
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
contextlib2==0.5.5
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-admin-rangefilter==0.5.2
django-cors-headers==3.0.2
django-debug-toolbar==2.0
django-envie==1.0.0
django-extensions==2.1.9
django-postgrespool2==1.0.0
django-filter==2.1.0
django-redis==4.10.0
django-rest-swagger==2.2.0
django-storages==1.7.1
django-templated-mail==1.1.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
djoser==2.0.1
docutils==0.15.2
entrypoints==0.3
facebook-sdk==3.1.0
google==2.0.3
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
gunicorn==19.9.0
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
Markdown==3.1.1
MarkupSafe==1.1.1
oauthlib==3.0.1
Pillow==6.1.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
pycrypto==2.6.1
pycryptodome==3.8.2
PyJWT==1.7.1
PySocks==1.7.0
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
PyYAML==5.1.1
raven==6.10.0
redis==3.2.1
requests==2.22.0
s3transfer==0.2.1
schema==0.7.1
sendgrid==6.1.0
sentry-sdk==0.10.2
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
twitter==1.18.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
